+++ CHANGELOG.md
@@ -4,6 +4,20 @@
 
 * `Realm.compactRealm()` works for encrypted Realms.
 *  Added `first(E defaultValue)` and `last(E defaultValue)` methods to `RealmList` and `RealmResult`. These methods will return the provided object instead of throwing an `IndexOutOfBoundsException` if the list is empty.
+*  Added `User.all()` that returns all known Realm Object Server users.
+
+## 2.0.3
+
+### Bug fixes
+
+* Those were not kept by ProGuard: names of native methods not in the `io.realm.internal` package, names of classes used in method signature (#3596).
+* Missing ProGuard configuration for libraries used by Sync extension (#3596).
+* Error handler was not called when sync session failed (#3597).
+
+### Enhancements
+
+* Reduce transformer logger verbosity (#3608).
+>>>>>>> base/master
 
 ## 2.0.2